formatting

+++ utils/balance_chars.c
@@ -1,4 +1,5 @@
-void balance_chars(char c, short* scale, char plus, char minus)
+void balance_chars(char c, int* scale,
+                   char plus, char minus)
 {
     if (c == plus) {
         (*scale)++;
